+++ app/src/main/java/com/omouravictor/invest_view/util/LocaleUtil.kt
@@ -19,7 +19,7 @@ object LocaleUtil {
         return intNumberFormat.format(value)
     }
 
-    fun getFormattedValueForPercent(value: Double): String {
+    fun getFormattedValueForPercent(value: Float): String {
         val percentFormat = NumberFormat.getPercentInstance(appLocale)
         percentFormat.maximumFractionDigits = 2
         return percentFormat.format(value)
